+++ examples/api2x/hello-decode/src/hello-decode.cpp
@@ -21,9 +21,8 @@
 void Usage(void) {
     printf("\n");
     printf("   Usage  :  hello-decode \n\n");
-    printf("     -sw/-hw        use software or hardware implementation\n");
     printf("     -i             input file name (HEVC elementary stream)\n\n");
-    printf("   Example:  hello-decode -sw  -i in.h265\n");
+    printf("   Example:  hello-decode -i in.h265\n");
     printf("   To view:  ffplay -f rawvideo -pixel_format yuv420p -video_size "
            "[width]x[height] %s\n\n",
            OUTPUT_FILE);
@@ -73,9 +72,9 @@ int main(int argc, char *argv[]) {
     // Implementation used must be the type requested from command line
     cfg[0] = MFXCreateConfig(loader);
     VERIFY(NULL != cfg[0], "MFXCreateConfig failed")
-
-    sts =
-        MFXSetConfigFilterProperty(cfg[0], (mfxU8 *)"mfxImplDescription.Impl", cliParams.implValue);
+    cfgVal[0].Type     = MFX_VARIANT_TYPE_U32;
+    cfgVal[0].Data.U32 = MFX_IMPL_TYPE_HARDWARE;
+    sts = MFXSetConfigFilterProperty(cfg[0], (mfxU8 *)"mfxImplDescription.Impl", cfgVal[0]);
     VERIFY(MFX_ERR_NONE == sts, "MFXSetConfigFilterProperty failed for Impl");
 
     // Implementation must provide an HEVC decoder