+++ CHANGELOG.md
@@ -12,6 +12,7 @@ portable media pipelines on CPUs, GPUs, and other accelerators.
 
 ### Removed 
 - USE_ONEAPI_INSTALL_LAYOUT build shortcut
+- '-sw' option from examples 
 
 ### Added
 - New MFXSetConfigFilterProperty property: `mfxExtendedDeviceId.RevisionID`